requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
neurokit2>=0.2.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
optuna[integration]>=3.0.0
shap>=0.42.0
imbalanced-learn>=0.11.0
ipywidgets>=8.1.0
ipython>=8.0.0
lightning>=2.1.0
torchmetrics>=1.0.0
plotly>=5.0.0
hydra-core>=1.3.2
wandb>=0.15.0
datasets>=2.14.0
streamlit>=1.25.0
deepchecks>=0.14.0
dvc>=3.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
# TimesFM (Time Series Foundation Model) — Google Research
timesfm[torch]>=1.3.0
huggingface-hub>=0.20.0
